fix size

+++ include/util/utils_inline.h
@@ -40,8 +40,8 @@ inline void get_size_bin_metadata(const std::string &bin_file, size_t &nrows,
             << ", ncols = " << ncols << std::endl;
 }
 
-inline void get_bin_metadata(const std::string &bin_file, size_t &nrows,
-                             size_t &ncols) {
+inline void get_bin_metadata(const std::string &bin_file, uint32_t &nrows,
+                             uint32_t &ncols) {
   std::ifstream reader(bin_file, std::ios::binary);
   reader.read((char *)&nrows, sizeof(uint32_t));
   reader.read((char *)&ncols, sizeof(uint32_t));
